Initial work on a DSL for our TEI schema instead of scribble/lp2.
The first goal is to have the contracts and the documentation
generated by the same syntactic forms, not just by the same file,
which will hopefully eliminate the need to manually keep things
in sync with the signature/unit system.

In the longer term, I hope to integrate defining (better) ADTs for the
elements into the DSL as well.