Version 0.4: Use custom #lang for our TEI schema.
From `ricoeur-portal`'s perspective, this is compatable
with version 0.3.1, though some documentation updates are needed.

In this version, the TEI XML contracts and documentation are
generated by the same syntactic forms, not just by the same file,
which eliminates the need to manually keep things
in sync via the signature/unit system.

The next goal for the `language-oriented` branch is integrate
defining (better) ADTs for the elements into `#lang ricoeur/tei/kernel`
as well, which will further reduce the number of things that
need to be manually kept in sync and hopefully provide more
useful and modular abstractions over the x-expressions
than the current class-based system.

The documentation in this version is correct, but not all of the added
things have been documented.