Version 0.4.1: Move old API to ricoeur/tei/oop. Remove ricoeur/lib.
As of this version, `ricoeur/tei` exports the new API for this
library (which is still incomplete).
Clients of the old API must switch to `ricoeur/tei/oop`, which
will be removed in a future release.

A corresponding commit must be made to `ricoeur-portal` before
this commit can be deployed.