Version 0.4.91: Patch for Windows line-breaks in `call/prettyprint-xml-out`.
On Windows, `xmllint` uses "\r\n" linebreaks in its output, even to stdout.
Because Racket (like C and most languages) internally uses "\n"
for line-breaks on all platforms, with conversion to platform-specific
endings isolated in specific IO functions, `call/prettyprint-xml-out`
must transform "\r\n" linebreaks before writing to `(current-output-port)`
to be usable in contexts like `with-output-to-string` and insertion into
GUI `text%` editor instances.

Failure to do this conversion is the likely cause of the inconsistency
in TEI Lint on Windows which impeded using the paragraph inference feature.

This commit does the conversion using a temporary file to take advantage
of `call-with-input-file*` with a `#:mode` argument of `'text`.
This might incur a performance penalty compared to writing directly
to a Racket pipe.
If so, I should explore whether the same conversion can be done on
non-file-stream ports (`reencode-input-port` seems to do something
slightly different).